Added first handin
First handin is about creating a few tables and populating them with data

This is done using python and sqlite, and basically it's just simple SQL.
Even tho the task was written in Swedish for use with Microsoft Access. For example, an INT that will be automatically incremented is reffered to as "Datatyp, räknare" and other things.